metabase: Move 'metabase' test fixture to a shared conftest
